Add cycleLength property to reminders
This createas a new property in the reminders model named cycleLength
which represents the time in hours that each reminder should be set to
every time that it's discarded. This also adds the discard endpoint
which deletes completely the reminder or creates a new one according to
the cycleLength property.